Treat build logs as binary files
Build logs can contain arbitrary binary data.  Although we generally
parse and display them as text, when fetching the log tail we may start
in the middle of a UTF-8 sequence, and in general we need to parse them
defensively.  Treat the files themselves as binary and decode them for
parsing purposes.