Skip tests involving daemonisation on Python 3 and Twisted < 16.4.0
Until Twisted 16.4.0, twistd hung indefinitely when run on Python 3.
Ubuntu 16.04 has Twisted 16.0.0, and backporting a newer version turns
out to have considerable complications; furthermore, we're going to need
to run at least some instances of launchpad-buildd on 16.04 for some
time, because we still need to run powerpc builders and the powerpc
architecture was removed in the 17.04 cycle.

Once launchpad-buildd supports both Python 2 and 3, the plan is to run
it on Python 3 when built for bionic and above, and otherwise run it on
Python 2.  To facilitate this, skip the one test that fails on Python 3
if we don't have a new enough Twisted.