Add a python3-lpbuildd binary package
This isn't used yet, but provides a basis for a full port on Ubuntu >=
18.04.

We now run tests on both Python 2 and 3 automatically.

The /usr/lib/launchpad-buildd/buildd-slave.tac symlink is now shipped by
launchpad-buildd rather than python-lpbuildd; it's only used when
running an actual builder rather than just using the libraries for
testing, and it would have caused python-lpbuildd and python3-lpbuildd
to conflict if we continued to ship it in the library packages.