Sanitize non-user-bound macaroons in build logs
Some of the macaroons issued by Launchpad are not bound to a user, for
example if they're scoped to a particular running build.  In those cases
the username part of the URL is empty.  Fix build log sanitization to
remove these macaroons.

This is typically not a significant security issue since the macaroons
in question cannot be used after the build has completed, and anyone who
can see a build while it's in progress should normally also be able to
see the resources it uses, but nevertheless it wasn't intended behaviour
and is an unnecessary exposure of information.