Add git_shallow_clone option to vcs_fetch() and use it for OCI build
Currently when cloning a git repository, the whole history is
cloned. That makes the cloning process slow (in case the repo is
large) and in some cases, the cloning fails due to its size (that's an
infrastructure problem but currently happens for OCI base images).
Just cloning the latest commit (via the "--depth 1" parameter)
improves the situation where large repos need to be cloned.

https://bugs.launchpad.net/launchpad-buildd/+bug/1939392