Refactor extra status handling to be common to all build types
Checking for the JSON status file on each status call is fairly cheap,
and it's convenient to have this in common code rather than having to
implement it separately for each build type that needs it.  I also made
the operation-side code a little more general so that it's easy to add
arbitrary key-value pairs to the current status, which will be useful
shortly.