Refactor multiple implementations of operation command helpers
We had several nearly-identical implementations of `run_build_command`
and related testtools matchers floating around.  Much of
launchpad-buildd is glue between other programs, so it makes sense for
it to have very careful checks on which subprocesses it runs, but it
doesn't need to have quite so much duplicated code in the process;
consolidate all these into common locations.