Add CI job support
Running jobs is handled via lpcraft.

The main complexity here is that we need to be able to return the output
of each individual lpcraft job separately, and so we need a novel
mechanism for returning the status of parts of a build farm job.  I
built this on top of the extra status file mechanism, previously used to
return VCS revision IDs.  This also means that we need a slightly more
complicated state machine than usual, since we have an initial
preparation step followed by iterating over a "run job" state for each
job.  Most of the rest of this is fairly typical for new build types.

At present there's a fair amount of overhead, since we're using standard
buildd containers and installing lpcraft in them, and then lpcraft
itself creates containers and sets them up according to
`.launchpad.yaml`.  We should be able to optimize this later by
providing pre-built containers.