Support CI pipeline stages with multiple jobs
The specification calls for either of these kinds of pipeline stage to
be valid:

    pipeline:
      # this stage is a list, which means jobs are executed in parallel
      - [test, lint]
      # this stage will only execute if previous steps in the pipeline
      # passed
      - build-wheel

In order to support this, we need to run all jobs in a stage (though not
yet in parallel) even if some of them fail, and only proceed to the next
stage if they all succeed.  We need to iterate over the pipeline in a
rather different way to support reasonable error reporting.

Note that this changes the format of the "jobs" argument passed to the
builder: it's now a list of lists of jobs, rather than a list of jobs.
This is OK since Launchpad hasn't started using it yet.