Avoid buggy PythonLoggingObserver on focal
The packaged version of Twisted in focal has a broken
`twisted.python.log.PythonLoggingObserver`
(https://bugs.launchpad.net/bugs/1961455).  While this should probably
be fixed in Ubuntu, it only affects one test suite, so it's easy to
avoid it by restructuring that suite using `FileLogObserver` from the
more modern `twisted.logger` package.

(Unfortunately we can't yet use the simpler
`twisted.logger.capturedLogs`, since the packaged version of Twisted in
focal is a little too old for that; but this construction is a similar
level of complexity to what we had before.)