Exclude bin/lpbuildd-git-proxy from pyupgrade
This is copied into the build chroot and run there, so it has different
compatibility constraints: it needs to continue supporting Python 3.5
for as long as we support things like building snaps on xenial.

Since my personal preferences for string formatting go f-strings >
%-formatting > `str.format`, I just reverted `pyupgrade`'s changes to
this file and excluded it from `pyupgrade` for the time being.