charm: Revamp build process and upgrade to focal
`charmcraft` works for reactive charms now, and seems to be simpler to
set up once you have the necessary `charmcraft.yaml` incantations.

I dropped support for manually pushing to the charm store, since once
this lands we should be able to set up a Launchpad recipe to push it to
Charmhub for us.