Fix gathering the output of charm recipe builds that use --build-path
Charm recipe builds have a "build path" option that allows configuring
them to run `charmcraft` in a subdirectory of the project rather than in
the project root.  This mostly worked, except that the build manager
tried to gather output files from the wrong place.  Fix it to gather
from the appropriate subdirectory in this case.