Convert daemon startup to systemd
As well as simplifying away 150 lines of shell script, this avoids
relying on su(1) to run the daemon as the `buildd` user, which in turn
avoids accidental pollution of the daemon's environment by whatever
su(1) thinks is appropriate in interactive environments.

There's a bit of extra complexity due to supporting multiple daemon
instances, which is mostly historical at this point, but it was easy
enough to handle using a template unit and a generator.

I dropped the explicit hostname condition, since all our builders
already have `RUN_NETWORK_REQUESTS_AS_ROOT=yes` set in
`/etc/default/launchpad-buildd`.

LP: #1964615