Ensure that launchpad-buildd runs with lxd as a supplementary group
`bin/sbuild-package` has an unfortunate hack to temporarily remove the
buildd user's lxd group membership (see
https://bugs.launchpad.net/launchpad-buildd/+bug/1820348).  If a
non-virtualized builder crashes for some reason in the middle of an
sbuild job, it will come back up without being reset to a clean image
(since it's non-virtualized), launchpad-buildd will start without the
lxd supplementary group it's supposed to have, and as a result it will
be unable to start LXD containers.

To work around all this, tell systemd to add lxd as a supplementary
group no matter what, ensuring that launchpad-buildd comes up properly
even if the system group database is temporarily wrong.