Add a timeout when revoking proxy tokens
We should have a timeout when making external requests so that builders
don't get tied up indefinitely if the external service in question
doesn't respond for whatever reason.  The builder proxy applies a
maximum lifetime to tokens, so the worst case of failing to revoke a
token is that it ends up being valid for a little longer than necessary.