Use a common output directory for all lpcraft jobs
In order to implement the `input` keyword in lpcraft, launchpad-buildd
needs to use the same top-level output directory for each job so that
lpcraft can find artifacts from previously-executed jobs.  This is
problematic with the current arrangements, because launchpad-buildd has
to pass different `--output-directory` options for each individual job
so that it can accurately determine which output artifacts belong to
which job.

However, with
https://code.launchpad.net/~cjwatson/lpcraft/+git/lpcraft/+merge/427724,
it becomes possible for launchpad-buildd to identify artifacts by job
name and index within a single common output directory, so we can take
advantage of that by looking for output files more precisely.

This also gathers the `properties` file (currently unused) in a slightly
different way, preparing for Launchpad to extract those output
properties and store them in the database in a way that can conveniently
be used downstream.