Work around hangs in Rust-based snap builds for armhf
LXCFS provides `/proc/cpuinfo` as a FUSE overlay inside containers to
filter visible CPUs based on cgroups.  On armhf, this results in the
wrong CPU features being visible to guests, because LXCFS itself is an
arm64 binary.  This causes Rust-based snap builds to hang on armhf,
because the combination of armhf userspace and a `/proc/cpuinfo` that
lacks the `neon` feature causes them to attempt to run ARMv6 code.

For now, work around this as suggested by the LXD maintainer by
unmounting `/proc/cpuinfo` inside the container.