Fix handling of librarian macaroons
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/429703
causes Launchpad to send authenticated librarian URLs to builders for
private source packages.  However, this needs a couple of fixes in
launchpad-buildd before it works.

Firstly, these macaroons are sent with an empty username, so we need to
send authentication if either the username or password is non-empty
rather than only if the username is non-empty.

Secondly, the librarian just sends 404 on authorization failures to
avoid revealing information, so we have to tell `urllib.request` to send
authentication credentials up-front (which is faster anyway) rather than
waiting for a 401.