Require the LXD snap to be installed
I don't love this since it's less declarative, but the lxd package no
longer exists in jammy, so we can't depend on it normally any more.  The
base OpenStack images where we typically install launchpad-buildd have
the lxd snap pre-installed, so this shouldn't cause any practical
inconvenience there.

This does mean that it's no longer possible to install launchpad-buildd
in a chroot, but it wouldn't have worked there for most purposes anyway.