Run lpcraft as the buildd user to allow nvidia.runtime=true to work
LXC's `nvidia` hook refuses to run if `both 0 0` is set in LXD's
`raw.idmap` key (mapping UID/GID 0 outside the container to UID/GID 0
inside).  I'm not exactly sure why that is, but running `lpcraft` as a
non-root user is fairly straightforward, so just do that to work around
it.