Don't require the lxd snap to be installed at postinst time
In commit 679f3234d4c21d12ca5469cbda9242d9fb27c647, I switched from
depending on the lxd .deb to requiring the snap to be installed, with
the aim of getting launchpad-buildd working on jammy.  Unfortunately,
this turned out to break upgrades of the riscv64 builders, where we
currently handle upgrading the base VM images using a chroot.

However, the only reason the postinst needed to care about this was so
that it could add the `buildd` user to the `lxd` group, and now that we
run launchpad-buildd from a systemd service we can safely just rely on
`SupplementaryGroups=lxd` giving the service that group membership.  Not
actually adding that user to that group avoids problems with the group
membership leaking through `sbuild`, allowing us to drop a chunk of
quite obscure code.