Disable OOM killing of launchpad-buildd
This came up as an idea at the last engineering sprint.  If a builder is
under memory pressure, we think that the kernel will sometimes choose to
kill the launchpad-buildd process, which is the worst thing it could
choose to do: the machine might as well be dead at that point.  We'd
always rather that it kill some other process so that we have a better
chance of getting a useful build log.

See systemd.exec(5).