snap: Add option to disable proxy after pull phase
We're still working on all the details, but this is a relatively simple
way to let us lock down certain snap recipe builds a bit more.

I converted the existing revocation code to use `requests`, since that's
been on my to-do list for a long time.  There's a minor testing gap
because the version of `responses` in Ubuntu 20.04 doesn't allow us a
way to inspect the `timeout` argument passed when making the request,
but we'll be able to fix that once we upgrade builders to Ubuntu 22.04,
which is in the works anyway.