Ignore fewer files when building recipes
`dpkg-buildpackage -I` ignores files matching a number of patterns (see
`@tar_ignore_default_pattern` in
`/usr/share/perl5/Dpkg/Source/Package.pm`).  Most of these are revision
control metadata, and ignoring those is no problem.  However, they also
include things like `.a` and `.so`.  Including these in source packages
is typically not good practice, but in the rare cases where people need
to do it, `dpkg-buildpackage -I` leaves them with no good alternative:
there's no way to override it and say that particular files shouldn't be
ignored.

Since by definition we're building from a clean bzr or git tree here,
the only things we actually need to ignore are the `.bzr` and `.git`
directories.  For everything else, we can reasonably assume that if it's
checked into revision control then the developer most probably wants to
have it in the source package.