Delete apt cache before the fetch-service set up
The fetch-service needs to be aware of all the contents of the
ubuntu archives in order to validate the requests for installations of packages
that the application makes.
The way that we have implemented this now means deleting Apt's cache and call
apt update right after the set up so that all of Apt's requests go through the fetch-service.