Enable fetch service for rock builds
Install the fetch service proxy on rock builds.
The reason why we need to configure some extra environment variables
is that some tools will not accept the self-signed certificate
used by the fetch-service by default (like cargo), and some tools
need to behave in a specific way for the fetch-service to work (like go).