feat: add HTTP_PROXY/HTTPS_PROXY to env
Proxy variable handling is all over the place on
Linux! We were already adding http_proxy and
https_proxy, but in quite a few cases I've needed
to special case this to then re-export the upper
case versions for tools to pick them up.

This commit adds the upper case equivalents.